Make sure to run all tests
