+++ packages/cli/src/util/config.ts
@@ -58,7 +58,7 @@ const CONFIG_FILE_SCHEMA: JSONSchemaType<FoundryConfig> = {
           oneOf: [
             {
               properties: {
-                "type": { enum: ["git-describe"], type: "string" },
+                "type": { const: "git-describe", type: "string" },
                 "tagPrefix": { type: "string", nullable: true },
               },
             },
